% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.combi.R
\name{add.combi}
\alias{add.combi}
\title{Add a genotyping array}
\usage{
add.combi(population, trait, combi.weights, trait.name = NULL)
}
\arguments{
\item{population}{population list}

\item{trait}{trait nr. for which to implement a combination of other traits}

\item{combi.weights}{Weights (only linear combinations of other traits are allowed!)}

\item{trait.name}{Name of the trait generated}
}
\value{
Population list
}
\description{
Function to add a genotyping array for the population
}
