% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.diag.R
\name{add.diag}
\alias{add.diag}
\title{Add something to the diagonal}
\usage{
add.diag(M, d)
}
\arguments{
\item{M}{Matrix}

\item{d}{Vector to add to the diagonal of the matrix}
}
\value{
Matrix with increased diagonal elements

Matrix with modified diagonal entries
}
\description{
Function to add numeric to the diagonal of a matrix
}
\examples{
A <- matrix(c(1,2,3,4), ncol=2)
B <- add.diag(A, 5)
}
