% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.founder.kinship.R
\name{add.founder.kinship}
\alias{add.founder.kinship}
\title{Add a genotyping array}
\usage{
add.founder.kinship(population, founder.kinship = "vanRaden", gen = 1)
}
\arguments{
\item{population}{population list}

\item{founder.kinship}{Default is to use vanRaden relationship. Alternative is to enter a pedigree-matrix (order of individuals is first male then female)}

\item{gen}{Generation for which to enter the pedigree-matrix}
}
\value{
Population list
}
\description{
Function to add a genotyping array for the population
}
\examples{
1+1 # to.be.entered
}
