% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.freq.R
\name{allele.freq}
\alias{allele.freq}
\title{Calculate allele frequcenies of a miraculix Z matrix without miraculix::allele_freq()}
\usage{
allele.freq(Z.code)
}
\arguments{
\item{Z.code}{bit-wise coded genotype matrix}
}
\description{
Calculate allele frequcenies of a miraculix Z matrix without miraculix::allele_freq()
}
