% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.bv.R
\name{analyze.bv}
\alias{analyze.bv}
\title{Analyze genomic values}
\usage{
analyze.bv(population, gen = NULL, database = NULL, cohorts = NULL,
  bvrow = "all", advanced = FALSE)
}
\arguments{
\item{population}{Population list}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{database}{Groups of individuals to consider for the export}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{bvrow}{Which traits to display (for multiple traits separte plots (par(mfrow)))}

\item{advanced}{Set to TRUE to also look at offspring pheno}
}
\value{
[[1]] Correlation between BV/BVE/Phenotypes [[2]] Genetic variance of the traits
}
\description{
Function to analyze correlation between bv/bve/pheno
}
\examples{
data(ex_pop)
analyze.bv(ex_pop,gen=1)
}
