% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwise.storing.R
\name{bit.snps}
\alias{bit.snps}
\title{Decoding of bitwise-storing in R}
\usage{
bit.snps(bit.seq, nbits, population = NULL, from.p.bit = 1)
}
\arguments{
\item{bit.seq}{bitwise stored SNP sequence}

\item{nbits}{Number of usable bits (default: 30)}

\item{population}{Population list}

\item{from.p.bit}{Bit to start on}
}
\value{
De-coded marker sequence
}
\description{
Function for decoding in bitwise-storing in R (only 30 of 32 bits are used!)
}
