% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.dataset.R
\name{cattle_chip}
\alias{cattle_chip}
\title{Cattle chip}
\source{
Ma et al 2015
}
\usage{
cattle_chip
}
\description{
Genome for cattle according to Ma et al.
}
\author{
Torsten Pook \email{torsten.pook@wur.nl}
}
