% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.up.R
\name{clean.up}
\alias{clean.up}
\title{Clean-up recombination points}
\usage{
clean.up(population, gen = "all", database = NULL, cohorts = NULL)
}
\arguments{
\item{population}{Population list}

\item{gen}{Generations to clean up (default: "current")}

\item{database}{Groups of individuals to consider}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}
}
\value{
Population-list with deleted irrelevant recombination points
}
\description{
Function to remove recombination points + origins with no influence on markers
}
\examples{
data(ex_pop)
ex_pop <- clean.up(ex_pop)
}
