% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeOriginsR.R
\name{decodeOriginsR}
\alias{decodeOriginsR}
\title{Origins-Decoding(R)}
\usage{
decodeOriginsR(P, row)
}
\arguments{
\item{P}{coded origins vector}

\item{row}{row to decode}
}
\value{
de-coded origins
}
\description{
R-Version of the internal bitwise-decoding of origins
}
\examples{
decodeOriginsR(0L)
}
