% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.mobps.R
\name{diag.mobps}
\alias{diag.mobps}
\title{Add a genotyping array}
\usage{
diag.mobps(elements)
}
\arguments{
\item{elements}{vector with entries to put on the diagonal of a matrix}
}
\value{
Diagonal matrix
}
\description{
Function to add a genotyping array for the population
}
\examples{
diag.mobps(5)
}
