% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.fromto.R
\name{edges.fromto}
\alias{edges.fromto}
\title{Detection of parental/child nodes}
\usage{
edges.fromto(edges)
}
\arguments{
\item{edges}{Edges of the json-file generated via the web-interface}
}
\value{
Matrix of Parent/Child-nodes for the considered edges
}
\description{
Internal function to extract parental/child node of an edge
}
