% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective.size.R
\name{effective.size}
\alias{effective.size}
\title{Estimate effective population size}
\usage{
effective.size(ld, dist, n)
}
\arguments{
\item{ld}{ld between markers}

\item{dist}{distance between markers in Morgan}

\item{n}{Population size}
}
\value{
Estimated effective population size
}
\description{
Internal function to estimate the effective population size
}
