% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.cohorts.individual.R
\name{get.cohorts.individual}
\alias{get.cohorts.individual}
\title{Derive ID on an individual}
\usage{
get.cohorts.individual(
  population,
  database = NULL,
  gen = NULL,
  cohorts = NULL,
  use.id = TRUE,
  keep.order = FALSE
)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{use.id}{Set to TRUE to use MoBPS ids instead of Sex_Nr_Gen based names}

\item{keep.order}{To not change order of individuals when ids are provided (default: TRUE)}
}
\value{
Individual ID for in gen/database/cohorts selected individuals
}
\description{
Function to derive the internal ID given to each individual
}
\examples{
data(ex_pop)
get.cohorts.individual(ex_pop, gen=2)
}
