% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.computing.time.R
\name{get.computing.time}
\alias{get.computing.time}
\title{Derive computing time}
\usage{
get.computing.time(
  population,
  verbose = TRUE,
  extend = FALSE,
  per.call = FALSE
)
}
\arguments{
\item{population}{Population list}

\item{verbose}{Set to FALSE to not display any prints}

\item{extend}{Set to TRUE to return computing times with detailled overview on generation and BVE (default: FALSE)}

\item{per.call}{Set to TRUE to return computing times per call of breeding.diploid / creating.diploid() (default: FALSE)}
}
\value{
Computing times overview
}
\description{
Function to derive computing time required to generate input population list
}
\examples{
data(ex_pop)
get.computing.time(ex_pop)
}
