% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.full.database.R
\name{get.full.database}
\alias{get.full.database}
\title{Generation of a database with one row per individual}
\usage{
get.full.database(
  population,
  database = NULL,
  gen = NULL,
  cohorts = NULL,
  id = NULL
)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{id}{Individual IDs to search/collect in the database}
}
\description{
Function to generate a database with one row per individual
}
