% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.geno.time.R
\name{get.geno.time}
\alias{get.geno.time}
\title{Derive class}
\usage{
get.geno.time(
  population,
  database = NULL,
  gen = NULL,
  cohorts = NULL,
  use.id = FALSE,
  use.all.copy = TRUE
)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{use.id}{Set to TRUE to use MoBPS ids instead of Sex_Nr_Gen based names (default: FALSE)}

\item{use.all.copy}{Set to TRUE to extract phenotyping}
}
\value{
Class of in gen/database/cohorts selected individuals
}
\description{
Function to devide the class for each individual
}
\examples{
data(ex_pop)
get.class(ex_pop, gen=2)
}
