% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.R
\name{get.id}
\alias{get.id}
\title{Derive ID on an individual}
\usage{
get.id(
  population,
  database = NULL,
  gen = NULL,
  cohorts = NULL,
  id = NULL,
  use.id = TRUE,
  keep.order = FALSE
)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{id}{Individual IDs to search/collect in the database}

\item{use.id}{Set to TRUE to use MoBPS ids instead of Sex_Nr_Gen based names (default: TRUE)}

\item{keep.order}{To not change order of individuals when ids are provided (default: FALSE)}
}
\value{
Individual ID for in gen/database/cohorts selected individuals
}
\description{
Function to derive the internal ID given to each individual
}
\examples{
data(ex_pop)
get.id(ex_pop, gen=2)
}
