% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.maf.R
\name{get.maf}
\alias{get.maf}
\title{Calculate minor allele frequencies}
\usage{
get.maf(population, database = NULL, gen = NULL, cohorts = NULL)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}
}
\value{
Allele frequency of the alternative allele
}
\description{
Function to calculate minor allele frequencies
}
\examples{
data(ex_pop)
get.maf(ex_pop, gen = 1)
}
