% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ngen.R
\name{get.ngen}
\alias{get.ngen}
\title{Number of generations}
\usage{
get.ngen(population)
}
\arguments{
\item{population}{Population list}
}
\value{
Numeric value
}
\description{
Function to calculate the number of generations in the population list
}
\examples{
data(ex_pop)
get.ngen(ex_pop)
}
