% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nindi.R
\name{get.nindi}
\alias{get.nindi}
\title{Number of generations}
\usage{
get.nindi(
  population,
  database = NULL,
  gen = NULL,
  cohorts = NULL,
  extended = FALSE,
  count.copy = FALSE
)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{count.copy}{Set to TRUE to double count individuals if multiple copies of an individual are included in gen/database/cohorts}
}
\value{
Numeric value
}
\description{
Function to calculate the number of generations in the population list
}
\examples{
data(ex_pop)
get.nindi(ex_pop)
}
