% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.npheno.R
\name{get.npheno}
\alias{get.npheno}
\title{Export underlying number of observations per phenotype}
\usage{
get.npheno(
  population,
  database = NULL,
  gen = NULL,
  cohorts = NULL,
  id = NULL,
  use.all.copy = FALSE,
  use.id = TRUE
)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{id}{Individual IDs to search/collect in the database}

\item{use.all.copy}{Set to TRUE to extract phenotyping}

\item{use.id}{Set to TRUE to use MoBPS ids instead of Sex_Nr_Gen based names (default: TRUE)}
}
\value{
Phenotypes for in gen/database/cohorts selected individuals
}
\description{
Function to export the number of observation of each underlying phenotype
}
\examples{
data(ex_pop)
get.pheno(ex_pop, gen=2)
}
