% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ngen.R
\name{get.ntrait}
\alias{get.ntrait}
\title{Number of traits}
\usage{
get.ntrait(population)
}
\arguments{
\item{population}{Population list}
}
\value{
Numeric value
}
\description{
Function to calculate the number of traits in the population list
}
\examples{
data(ex_pop)
get.ngen(ex_pop)
}
