% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pheno.off.count.R
\name{get.pheno.off.count}
\alias{get.pheno.off.count}
\title{Export underlying number of used offspring for offspring phenotypes}
\usage{
get.pheno.off.count(
  population,
  database = NULL,
  gen = NULL,
  cohorts = NULL,
  id = NULL,
  use.id = TRUE
)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{id}{Individual IDs to search/collect in the database}

\item{use.id}{Set to TRUE to use MoBPS ids instead of Sex_Nr_Gen based names (default: TRUE)}
}
\value{
Number of offspring with phenotypes for in gen/database/cohorts selected individuals
}
\description{
Function to export number of observations used for offspring phenotypes
}
\examples{
data(ex_pop)
get.pheno.off.count(ex_pop, gen=2)
}
