% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.qtl.R
\name{get.qtl.effects}
\alias{get.qtl.effects}
\title{QTL effect extraction}
\usage{
get.qtl.effects(population)
}
\arguments{
\item{population}{Population list}
}
\value{
List with [[1]] single SNP QTLs [[2]] epistatic SNP QTLs [[3]] dice QTL
}
\description{
Function to extract QTL effect sizes
}
\examples{
data(ex_pop)
effects <- get.qtl.effects(ex_pop)
}
