% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ngen.R, R/get.size.R
\name{get.size}
\alias{get.size}
\title{Number of individuals in each generation}
\usage{
get.size(population)

get.size(population)
}
\arguments{
\item{population}{Population list}
}
\value{
matrix with numeric values

matrix with numeric values
}
\description{
Function to calculate the number of individuals per generation

Function to calculate the number of individuals per generation
}
\examples{
data(ex_pop)
get.size(ex_pop)
data(ex_pop)
get.size(ex_pop)
}
