% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ngen.R
\name{get.trait.name}
\alias{get.trait.name}
\title{Number of generations}
\usage{
get.trait.name(population)
}
\arguments{
\item{population}{Population list}
}
\value{
Numeric value
}
\description{
Function to export trait names in the population list
}
\examples{
data(ex_pop)
get.ngen(ex_pop)
}
