% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.variance.components.R
\name{get.variance.components}
\alias{get.variance.components}
\title{Derive variance components}
\usage{
get.variance.components(population)
}
\arguments{
\item{population}{Population list}
}
\value{
Estimated variance components
}
\description{
Function to derive variance components
}
\examples{
data(ex_pop)
population = breeding.diploid(ex_pop, bve = TRUE, bve.gen = 2)
get.variance.components(population)
}
