% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship.development.R
\name{kinship.development}
\alias{kinship.development}
\title{Devolopment of genetic/breeding value}
\usage{
kinship.development(population, database = NULL, gen = NULL,
  cohorts = NULL, json = FALSE, ibd.obs = 50, hbd.obs = 10)
}
\arguments{
\item{population}{population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{json}{If TRUE extract which cohorts to plot according to the json-file used in json.simulation}

\item{ibd.obs}{Number of Individual pairs to sample for IBD estimation}

\item{hbd.obs}{Number of Individuals to sample for HBD estimation}
}
\description{
Function to plot genetic/breeding values for multiple generation/cohorts
}
