% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.dataset.R
\name{maize_chip}
\alias{maize_chip}
\title{maize chip}
\source{
Lee et al 2002
}
\usage{
maize_chip
}
\description{
Genome for maize according to Lee et al.
}
\author{
Torsten Pook \email{torsten.pook@wur.nl}
}
