% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.cohorts.R
\name{merge.cohorts}
\alias{merge.cohorts}
\title{Merging of cohorts}
\usage{
\method{merge}{cohorts}(population, cohorts, name.cohort = NULL)
}
\arguments{
\item{population}{Population list}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{name.cohort}{Name of the newly added cohort}
}
\description{
Function to merge cohorts in a population list
}
