% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miraculix.R
\name{miraculix}
\alias{miraculix}
\title{Add miraculix-coding for genotypes}
\usage{
miraculix(population)
}
\arguments{
\item{population}{Population list}
}
\value{
Population list
}
\description{
Internal function to store genotypes bit-wise
}
