% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGC.R
\name{OGC}
\alias{OGC}
\title{Optimal genetic contribution}
\usage{
OGC(A, u, Q, cAc = NA, single = TRUE)
}
\arguments{
\item{A}{relationship matrix}

\item{u}{breeding values}

\item{Q}{sex indicator}

\item{cAc}{target gain in inbreeding}

\item{single}{If FALSE multiple individuals can be removed at the same type (this is faster but potentially inaccurate!)}
}
\description{
In this function the OGC selection according to Meuwissen 1997 is performed
}
