% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.cores.R
\name{optimize.cores}
\alias{optimize.cores}
\title{Optimize generation cores}
\usage{
optimize.cores(
  population = NULL,
  test.size = 2500,
  max.cores = 10,
  verbose = TRUE,
  plot = TRUE
)
}
\arguments{
\item{population}{Population list}

\item{test.size}{Number of individuals to generate for each core number (default: 2500)}

\item{max.cores}{Maximum number of cores to test (default: 10)}

\item{verbose}{Set to FALSE to not display any prints}

\item{plot}{Set to FALSE to not generate a plot of computing times per core}
}
\value{
Population-list with one or more additional new traits
}
\description{
Function to optimize the number of cores used in the generation of new individuals
}
\examples{
population = optimize.cores(max.cores=2, test.size=500)
}
