% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedmap.to.phasedbeaglevcf.R
\name{pedmap.to.phasedbeaglevcf}
\alias{pedmap.to.phasedbeaglevcf}
\title{Perform imputing/phasing}
\usage{
pedmap.to.phasedbeaglevcf(ped_path = NULL, map_path = NULL,
  beagle_jar = "beagle.03Jul18.40b.jar", plink_dir = "./Plink/plink",
  db_dir = "Plink/DB/")
}
\arguments{
\item{ped_path}{Directory of the ped-file}

\item{map_path}{Directory of the map-file}

\item{beagle_jar}{Directory of BEAGLE}

\item{plink_dir}{Directory of plink}

\item{db_dir}{Directory to save newly generated files (ped/map will be stored in the original folder)}
}
\description{
Perform imputing/phasing (path chosen for the web-based application)
}
