% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.population.R
\name{plot.population}
\alias{plot.population}
\title{Plot Population}
\usage{
\method{plot}{population}(x, type = "bve", gen = NULL, database = NULL, cohorts = NULL, ...)
}
\arguments{
\item{x}{Population-list}

\item{type}{Default "bve" - bv.development, alt: "kinship" - kinship.development(), "pca" - get.pca()}

\item{gen}{generations to consider}

\item{database}{groups to consider}

\item{cohorts}{cohorts to consider}

\item{...}{remaining stuff}
}
\value{
Summary of the population list including number of individuals, genome length and trait overview
}
\description{
Basic plot of the population list
}
\examples{
data(ex_pop)
plot(ex_pop)
}
