% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.function.haldane.R
\name{recombination.function.haldane}
\alias{recombination.function.haldane}
\title{Haldane recombination function}
\usage{
recombination.function.haldane(noc, length.genome)
}
\arguments{
\item{noc}{number of recombination points}

\item{length.genome}{Length of the genome}
}
\description{
Standard haldane function
}
