% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.cohort.R
\name{rename.cohort}
\alias{rename.cohort}
\title{Function to exclude individuals from a database}
\usage{
rename.cohort(population, old.name, new.name, verbose = TRUE)
}
\arguments{
\item{population}{Population list}

\item{old.name}{Quick-insert for database (vector of names of cohorts to export)}

\item{new.name}{Groups of individuals to remove from the database (same IDs!)}
}
\value{
population list
}
\description{
Function to exclude individuals from a database
}
\examples{
data(ex_pop)
population <- rename.cohort(ex_pop, old.name="Cohort_4", new.name = "NewName")
}
