% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.cohort.R
\name{renaming.cohort}
\alias{renaming.cohort}
\title{Rename a cohort}
\usage{
renaming.cohort(population, old.name, new.name, verbose = TRUE)
}
\arguments{
\item{population}{Population list}

\item{old.name}{Quick-insert for database (vector of names of cohorts to export)}

\item{new.name}{Groups of individuals to remove from the database (same IDs!)}

\item{verbose}{Set to FALSE to not display any prints}
}
\value{
population list
}
\description{
Function to rename a cohort
}
\examples{
data(ex_pop)
population <- renaming.cohort(ex_pop, old.name="Cohort_4", new.name = "NewName")
}
