% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMedian.R
\name{rowMedian}
\alias{rowMedian}
\title{Row-wise Median}
\usage{
rowMedian(A)
}
\arguments{
\item{A}{Matrix}
}
\value{
row-wise median values of a matrix

Matrix with modified diagonal entries
}
\description{
Function to calculate row-wise median values
}
\examples{
A <- matrix(c(1,2,3,4), ncol=2)
x <- rowMedian(A)
}
