% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.relationship.R
\name{scaling.relationship}
\alias{scaling.relationship}
\title{scaling.relationship}
\usage{
scaling.relationship(A, Z, p)
}
\arguments{
\item{A}{Population list}

\item{Z}{genotype matrix}

\item{p}{allele frequency}
}
\value{
scaled genomic relationship matrix
}
\description{
Function to scale relationship matricies
}
