% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.age.point.R
\name{set.age.point}
\alias{set.age.point}
\title{Set age points}
\usage{
set.age.point(
  population,
  database = NULL,
  gen = NULL,
  cohorts = NULL,
  time.point = 0
)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{time.point}{Input value for the age point (time of birth) of an individual (default: 0)}
}
\value{
Population-List with newly entered class values
}
\description{
Function to overwrite age.points of individuals
}
\examples{
data(ex_pop)
population <- set.age.point(ex_pop, database=cbind(1,1), time.point = 2)
}
