% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.pedigree.mixblup.R
\name{write.pedigree}
\alias{write.pedigree}
\title{write.pedigree.mixblup}
\usage{
write.pedigree(
  population,
  path,
  gen = NULL,
  database = NULL,
  cohorts = NULL,
  id = NULL,
  depth.pedigree = 7,
  storage.save = 1.5,
  verbose = TRUE,
  mixblup.reliability = FALSE,
  blupf90 = FALSE,
  include.error = TRUE
)
}
\arguments{
\item{population}{Population list}

\item{path}{AA}

\item{gen}{AA}

\item{database}{AA}

\item{cohorts}{AA}

\item{id}{AA}

\item{depth.pedigree}{AA}

\item{storage.save}{AA}

\item{verbose}{AA}

\item{mixblup.reliability}{AA}

\item{blupf90}{FALSE for mixblup; TRUE for MixBLUP}

\item{include.error}{AA}
}
\value{
pedigree table
}
\description{
write.pedigree.mixblup
}
