% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.combi.R
\name{add.combi}
\alias{add.combi}
\title{Add a trait as a linear combination of other traits}
\usage{
add.combi(population, trait, combi.weights, trait.name = NULL)
}
\arguments{
\item{population}{population list}

\item{trait}{trait nr. for which to implement a combination of other traits}

\item{combi.weights}{Weights (only linear combinations of other traits are allowed!)}

\item{trait.name}{Name of the trait generated}
}
\value{
Population list

Population list
}
\description{
Function to create an additional trait that is the results of a linear combination of the other traits
}
\examples{
data(ex_pop)
population <- creating.trait(ex_pop, n.additive = c(100,100), replace.traits = TRUE)
population <- add.combi(population, trait = 3, combi.weights = c(1,2))
}
