% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.founder.kinship.R
\name{add.founder.kinship}
\alias{add.founder.kinship}
\title{Add a relationship matrix for founder individuals}
\usage{
add.founder.kinship(population, founder.kinship = "GBLUP", gen = 1)
}
\arguments{
\item{population}{population list}

\item{founder.kinship}{Default is to use vanRaden relationship. Alternative is to enter a pedigree-matrix (order of individuals is first male then female)}

\item{gen}{Generation for which to enter the pedigree-matrix}
}
\value{
Population list
}
\description{
Function to relationship matrix for founder individuals that is used for any calculation of the pedigree
}
\examples{
data(ex_pop)
population <- add.founder.kinship(ex_pop)
}
