% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demiraculix.R
\name{demiraculix}
\alias{demiraculix}
\title{Remove miraculix-coding for genotypes}
\usage{
demiraculix(population)
}
\arguments{
\item{population}{Population list}
}
\value{
Population list
}
\description{
Internal function to decode all genotypes to non-miraculix objects
}
\examples{
# This is only relevant with the package miraculix is installed and used
population <- creating.diploid(nsnp=100, nindi=50)
\donttest{population <- demiraculix(population)}
}
