% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.individual.R
\name{generation.individual}
\alias{generation.individual}
\title{Function to generate a new individual}
\usage{
generation.individual(
  indexb,
  population,
  info_father_list,
  info_mother_list,
  copy.individual,
  mutation.rate,
  remutation.rate,
  recombination.rate,
  recom.f.indicator,
  recom.f.polynom,
  duplication.rate,
  duplication.length,
  duplication.recombination,
  delete.same.origin,
  gene.editing,
  nr.edits,
  gen.architecture.m,
  gen.architecture.f,
  decodeOriginsU,
  current.gen,
  save.recombination.history,
  new.bv.child,
  dh.mating,
  share.genotyped,
  added.genotyped,
  genotyped.array,
  dh.sex,
  n.observation
)
}
\arguments{
\item{indexb}{windows parallel internal test}

\item{population}{windows parallel internal test}

\item{info_father_list}{windows parallel internal test}

\item{info_mother_list}{windows parallel internal test}

\item{copy.individual}{windows parallel internal test}

\item{mutation.rate}{windows parallel internal test}

\item{remutation.rate}{windows parallel internal test}

\item{recombination.rate}{windows parallel internal test}

\item{recom.f.indicator}{windows parallel internal test}

\item{recom.f.polynom}{windows parallel internal test}

\item{duplication.rate}{windows parallel internal test}

\item{duplication.length}{windows parallel internal test}

\item{duplication.recombination}{windows parallel internal test}

\item{delete.same.origin}{windows parallel internal test}

\item{gene.editing}{windows parallel internal test}

\item{nr.edits}{windows parallel internal test}

\item{gen.architecture.m}{windows parallel internal test}

\item{gen.architecture.f}{windows parallel internal test}

\item{decodeOriginsU}{windows parallel internal test}

\item{current.gen}{windows parallel internal test}

\item{save.recombination.history}{windows parallel internal test}

\item{new.bv.child}{windows parallel internal test}

\item{dh.mating}{windows parallel internal test}

\item{share.genotyped}{windows parallel internal test}

\item{added.genotyped}{windows parallel internal test}

\item{genotyped.array}{windows parallel internal test}

\item{dh.sex}{windows parallel internal test}

\item{n.observation}{windows parallel internal test}
}
\value{
Offspring individual
}
\description{
Function to generate a new individual
}
