% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.database.R
\name{get.database}
\alias{get.database}
\title{gen/database/cohorts conversion}
\usage{
get.database(
  population,
  gen = NULL,
  database = NULL,
  cohorts = NULL,
  avoid.merging = FALSE
)
}
\arguments{
\item{population}{Population list}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{database}{Groups of individuals to consider for the export}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{avoid.merging}{Set to TRUE to avoid different cohorts to be merged in a joint group when possible}
}
\value{
Combine gen/database/cohorts to a joined database
}
\description{
Function to derive a database based on gen/database/cohorts
}
\examples{
data(ex_pop)
get.database(ex_pop, gen=2)
}
