% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.genotyped.R
\name{get.genotyped}
\alias{get.genotyped}
\title{Derive genotyping status}
\usage{
get.genotyped(
  population,
  database = NULL,
  gen = NULL,
  cohorts = NULL,
  use.id = TRUE,
  use.all.copy = FALSE
)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{use.id}{Set to TRUE to use MoBPS ids instead of Sex_Nr_Gen based names (default: TRUE)}

\item{use.all.copy}{Set to TRUE to extract phenotyping}
}
\value{
Check if in gen/database/cohorts selected individuals are genotyped
}
\description{
Function to if selected individuals are genotyped
}
\examples{
data(ex_pop)
get.genotyped(ex_pop, gen=2)
}
