% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.grm.R
\name{get.grm}
\alias{get.grm}
\title{Derive genotypes of selected individuals}
\usage{
get.grm(
  population,
  geno = NULL,
  database = NULL,
  gen = NULL,
  cohorts = NULL,
  id = NULL,
  use.id = TRUE,
  array = NULL,
  p.database = NULL,
  p.gen = NULL,
  p.cohorts = NULL,
  p.id = NULL
)
}
\arguments{
\item{population}{Population list}

\item{geno}{Manually provided genotype matrix (SNP x Indi)}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{id}{Individual IDs to search/collect in the database}

\item{use.id}{Set to TRUE to use MoBPS ids instead of Sex_Nr_Gen based names (default: TRUE)}

\item{array}{Use only markers available on the array}

\item{p.database}{Groups of individuals to consider for the export}

\item{p.gen}{Quick-insert for database (vector of all generations to export)}

\item{p.cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{p.id}{Individual IDs to search/collect in the database}
}
\value{
Genotype data for in gen/database/cohorts selected individuals
}
\description{
Function to derive genotypes of selected individuals
}
\examples{
data(ex_pop)
G <- get.grm(ex_pop, gen=2)
}
