% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.R
\name{get.id}
\alias{get.id}
\title{Derive ID on an individual}
\usage{
get.id(population, database = NULL, gen = NULL, cohorts = NULL)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}
}
\value{
Individual ID for in gen/database/cohorts selected individuals
}
\description{
Function to derive the internal ID given to each individual
}
\examples{
data(ex_pop)
get.id(ex_pop, gen=2)
}
