% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.infos.R
\name{get.infos}
\alias{get.infos}
\title{Extract bv/pheno/geno of selected individuals}
\usage{
get.infos(population, database = NULL, gen = NULL, cohorts = NULL)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}
}
\value{
Info list [[1]] phenotypes [[2]] genomic values [[3]] Z [[4/5/6]] additive/epistatic/dice marker effects
}
\description{
Function to extract bv/pheno/geno of selected individuals
}
\examples{
data(ex_pop)
get.infos(ex_pop, gen=2)
}
