% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ngen.R
\name{get.ngen}
\alias{get.ngen}
\title{Number of generations}
\usage{
get.ngen(population)
}
\arguments{
\item{population}{Population list}
}
\value{
Population-list with one or more additional new traits
}
\description{
Set default parameter values in breeding.diploid
}
\examples{
data(ex_pop)
get.ngen(ex_pop)
}
