% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.variance.components.R
\name{get.variance.components}
\alias{get.variance.components}
\title{Derive age point}
\usage{
get.variance.components(population)
}
\arguments{
\item{population}{Population list}
}
\value{
Time point selected gen/database/cohorts-individuals are born
}
\description{
Function to devide age point for each individual (Same as time.point unless copy.individual is used for aging)
}
\examples{
data(ex_pop)
get.age.point(ex_pop, gen=2)
}
