% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.simulation.R
\name{json.simulation}
\alias{json.simulation}
\title{Generation of the starting population}
\usage{
json.simulation(file = NULL, total = NULL)
}
\arguments{
\item{file}{Path to a json-file generated by the user-interface}

\item{total}{Json-file imported via jsonlite::read_json}
}
\description{
Generation of the starting population
}
