% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.posdef.R
\name{matrix.posdef}
\alias{matrix.posdef}
\title{Projection positive definit}
\usage{
matrix.posdef(A, verbose = TRUE, matrix.name = "Matrix", epsilon = 0.0000001)
}
\arguments{
\item{A}{Input matrix to project}

\item{verbose}{Set to FALSE to not display any prints}

\item{matrix.name}{This is just for internal prints}

\item{epsilon}{This factor is added to the diagonal to avoid numerical issues with semi-definit matrices}
}
\value{
Positive definit matrix
}
\description{
Function to project a matrix in the space of positive definit matrices
}
