% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.cohorts.R
\name{merging.cohorts}
\alias{merging.cohorts}
\title{Merging of cohorts}
\usage{
merging.cohorts(population, cohorts, name.cohort = NULL)
}
\arguments{
\item{population}{Population list}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{name.cohort}{Name of the newly added cohort}
}
\description{
Function to merge cohorts in a population list
}
\examples{
data(ex_pop)
population <- breeding.diploid(ex_pop, add.gen = 7, breeding.size = 50)
population <- breeding.diploid(population, add.gen = 7, breeding.size = 50,
 selection.m.database = cbind(6,1), selection.f.database = cbind(6,2))
population <- merging.cohorts(population, cohorts = c("Cohort_7_M", "Cohort_8_M"))
}
