% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.intro.R
\name{mutation.intro}
\alias{mutation.intro}
\title{Mutation intro}
\usage{
mutation.intro(
  population,
  gen = NULL,
  database = NULL,
  cohorts = NULL,
  qtl.posi,
  target.variant = NULL,
  haplo.set = 1
)
}
\arguments{
\item{population}{Population list}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{database}{Groups of individuals to consider for the export}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}

\item{qtl.posi}{Marker number to mutate}

\item{target.variant}{target variant to obtain (( if haplotype already is correct do not introduce a mutation ))}

\item{haplo.set}{Select chromosome set (default: 1 , alt: 2, 1:2 (to edit both))}
}
\value{
Population-List with mutated marker for the selected individual
}
\description{
Function to change the base-pair in a specific loci
}
\examples{
data(ex_pop)
ex_pop <- mutation.intro(ex_pop, database = cbind(1,1,1), qtl.posi=100)
}
