% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.relationship.R
\name{scaling.relationship}
\alias{scaling.relationship}
\title{scaling.relationship}
\usage{
scaling.relationship(A, Z, p)
}
\arguments{
\item{A}{Population list}

\item{Z}{AA}

\item{p}{AA}
}
\value{
scaled genomic relationship matrix
}
\description{
scaling.relationship
}
\examples{
write.pedigree.mixblup
}
