% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.pedigree.mixblup.R
\name{write.pedigree.mixblup}
\alias{write.pedigree.mixblup}
\title{write.pedigree.mixblup}
\usage{
write.pedigree.mixblup(
  population,
  path,
  gen = NULL,
  database = NULL,
  cohorts = NULL,
  id = NULL,
  depth.pedigree = 7,
  storage.save = 1.5,
  verbose = TRUE
)
}
\arguments{
\item{population}{Population list}
}
\value{
write.pedigree.mixblup
}
\description{
write.pedigree.mixblup
}
\examples{
write.pedigree.mixblup
}
